[#366] Split db commands
