+++ recommender/commands/__init__.py
@@ -3,3 +3,4 @@
 from .db import db_command
 from .migrate import migrate_command
 from .train import train_command
+from .seed import seed_command
